+++ EyeTrackApp/camera_widget.py
@@ -10,9 +10,11 @@ from camera import Camera, CameraState
 from osc import EyeId
 import cv2
 import sys
-from utils.misc_utils import PlaySound, SND_FILENAME,SND_ASYNC, resource_path
+from utils.misc_utils import PlaySound, SND_FILENAME, SND_ASYNC, resource_path
 import traceback
 import numpy as np
+
+
 class CameraWidget:
     def __init__(self, widget_id: EyeId, main_config: EyeTrackConfig, osc_queue: Queue):
         self.gui_camera_addr = f"-CAMERAADDR{widget_id}-"
@@ -46,7 +48,9 @@ class CameraWidget:
         elif self.eye_id == EyeId.LEFT:
             self.config = main_config.left_eye
         else:
-            raise RuntimeError("\033[91m[WARN] Cannot have a camera widget represent both eyes!\033[0m")
+            raise RuntimeError(
+                "\033[91m[WARN] Cannot have a camera widget represent both eyes!\033[0m"
+            )
 
         self.cancellation_event = Event()
         # Set the event until start is called, otherwise we can block if shutdown is called.
@@ -80,14 +84,14 @@ class CameraWidget:
 
         self.roi_layout = [
             [
-                sg.Graph( 
+                sg.Graph(
                     (640, 480),
                     (0, 480),
                     (640, 0),
                     key=self.gui_roi_selection,
                     drag_submits=True,
                     enable_events=True,
-                    background_color='#424042',
+                    background_color="#424042",
                 )
             ]
         ]
@@ -95,34 +99,50 @@ class CameraWidget:
         # Define the window's contents
         self.tracking_layout = [
             [
-                sg.Text("Rotation", background_color='#424042'),
+                sg.Text("Rotation", background_color="#424042"),
                 sg.Slider(
                     range=(0, 360),
                     default_value=self.config.rotation_angle,
                     orientation="h",
                     key=self.gui_rotation_slider,
-                    background_color='#424042',
-                    tooltip = "Adjust the rotation of your cameras, make them level.",
+                    background_color="#424042",
+                    tooltip="Adjust the rotation of your cameras, make them level.",
                 ),
             ],
             [
-                sg.Button("Start Calibration", key=self.gui_restart_calibration, button_color='#6f4ca1', tooltip = "Start eye calibration. Look all arround to all extreams without blinking until sound is heard.",),
-                sg.Button("Stop Calibration", key=self.gui_stop_calibration, button_color='#6f4ca1', tooltip = "Stop eye calibration manualy.",),
-                sg.Button("Recenter Eyes", key=self.gui_recenter_eyes, button_color='#6f4ca1', tooltip = "Make your eyes center again.",),
-
+                sg.Button(
+                    "Start Calibration",
+                    key=self.gui_restart_calibration,
+                    button_color="#6f4ca1",
+                    tooltip="Start eye calibration. Look all arround to all extreams without blinking until sound is heard.",
+                ),
+                sg.Button(
+                    "Stop Calibration",
+                    key=self.gui_stop_calibration,
+                    button_color="#6f4ca1",
+                    tooltip="Stop eye calibration manualy.",
+                ),
+                sg.Button(
+                    "Recenter Eyes",
+                    key=self.gui_recenter_eyes,
+                    button_color="#6f4ca1",
+                    tooltip="Make your eyes center again.",
+                ),
             ],
             [
-                sg.Text("Mode:", background_color='#424042'),
-                sg.Text("Calibrating", key=self.gui_mode_readout, background_color='#424042'),
-                sg.Text("", key=self.gui_tracking_fps, background_color='#424042'),
-                sg.Text("", key=self.gui_tracking_bps, background_color='#424042'),
-            #    sg.Checkbox(
-            #        "Circle crop:",
-            #        default=self.config.gui_circular_crop,
-            #        key=self.gui_circular_crop,
-            #        background_color='#424042',
-            #        tooltip = "Circle crop only applies to RANSAC3D and Blob.",
-            #    ),
+                sg.Text("Mode:", background_color="#424042"),
+                sg.Text(
+                    "Calibrating", key=self.gui_mode_readout, background_color="#424042"
+                ),
+                sg.Text("", key=self.gui_tracking_fps, background_color="#424042"),
+                sg.Text("", key=self.gui_tracking_bps, background_color="#424042"),
+                #    sg.Checkbox(
+                #        "Circle crop:",
+                #        default=self.config.gui_circular_crop,
+                #        key=self.gui_circular_crop,
+                #        background_color='#424042',
+                #        tooltip = "Circle crop only applies to RANSAC3D and Blob.",
+                #    ),
             ],
             [sg.Image(filename="", key=self.gui_tracking_image)],
             [
@@ -135,25 +155,57 @@ class CameraWidget:
                     drag_submits=True,
                     enable_events=True,
                 ),
-                sg.Text("Please set an Eye Cropping.", key=self.gui_roi_message, background_color='#424042', visible=False),
+                sg.Text(
+                    "Please set an Eye Cropping.",
+                    key=self.gui_roi_message,
+                    background_color="#424042",
+                    visible=False,
+                ),
             ],
         ]
 
         self.widget_layout = [
             [
-                sg.Text("Camera Address", background_color='#424042'),
-                sg.InputText(self.config.capture_source, key=self.gui_camera_addr, tooltip = "Enter the IP address or UVC port of your camera. (Include the 'http://')",),
+                sg.Text("Camera Address", background_color="#424042"),
+                sg.InputText(
+                    self.config.capture_source,
+                    key=self.gui_camera_addr,
+                    tooltip="Enter the IP address or UVC port of your camera. (Include the 'http://')",
+                ),
             ],
             [
-                sg.Button("Save and Restart Tracking", key=self.gui_save_tracking_button, button_color='#6f4ca1'),
+                sg.Button(
+                    "Save and Restart Tracking",
+                    key=self.gui_save_tracking_button,
+                    button_color="#6f4ca1",
+                ),
             ],
             [
-                sg.Button("Tracking Mode", key=self.gui_tracking_button, button_color='#6f4ca1', tooltip = "Go here to track your eye.",),
-                sg.Button("Cropping Mode", key=self.gui_roi_button, button_color='#6f4ca1', tooltip = "Go here to crop out your eye.",),
+                sg.Button(
+                    "Tracking Mode",
+                    key=self.gui_tracking_button,
+                    button_color="#6f4ca1",
+                    tooltip="Go here to track your eye.",
+                ),
+                sg.Button(
+                    "Cropping Mode",
+                    key=self.gui_roi_button,
+                    button_color="#6f4ca1",
+                    tooltip="Go here to crop out your eye.",
+                ),
             ],
             [
-                sg.Column(self.tracking_layout, key=self.gui_tracking_layout, background_color='#424042'),
-                sg.Column(self.roi_layout, key=self.gui_roi_layout, background_color='#424042', visible=False),
+                sg.Column(
+                    self.tracking_layout,
+                    key=self.gui_tracking_layout,
+                    background_color="#424042",
+                ),
+                sg.Column(
+                    self.roi_layout,
+                    key=self.gui_roi_layout,
+                    background_color="#424042",
+                    visible=False,
+                ),
             ],
         ]
 
@@ -203,7 +255,11 @@ class CameraWidget:
             event == self.gui_save_tracking_button
             and values[self.gui_camera_addr] != self.config.capture_source
         ):
-            print("\033[94m[INFO] New value: {}\033[0m".format(values[self.gui_camera_addr]))
+            print(
+                "\033[94m[INFO] New value: {}\033[0m".format(
+                    values[self.gui_camera_addr]
+                )
+            )
             try:
                 # Try storing ints as ints, for those using wired cameras.
                 self.config.capture_source = int(values[self.gui_camera_addr])
@@ -211,20 +267,24 @@ class CameraWidget:
                 if values[self.gui_camera_addr] == "":
                     self.config.capture_source = None
                 else:
-                    if len(values[self.gui_camera_addr]) > 5 and "http" not in values[self.gui_camera_addr] and ".mp4" not in values[self.gui_camera_addr]: # If http is not in camera address, add it.
-                        self.config.capture_source = f"http://{values[self.gui_camera_addr]}/"   
+                    if (
+                        len(values[self.gui_camera_addr]) > 5
+                        and "http" not in values[self.gui_camera_addr]
+                        and ".mp4" not in values[self.gui_camera_addr]
+                    ):  # If http is not in camera address, add it.
+                        self.config.capture_source = (
+                            f"http://{values[self.gui_camera_addr]}/"
+                        )
                     else:
                         self.config.capture_source = values[self.gui_camera_addr]
             changed = True
-
-
 
         if self.config.rotation_angle != values[self.gui_rotation_slider]:
             self.config.rotation_angle = int(values[self.gui_rotation_slider])
             changed = True
 
-      # if self.config.gui_circular_crop != values[self.gui_circular_crop]:
-       #     self.config.gui_circular_crop = values[self.gui_circular_crop]
+        # if self.config.gui_circular_crop != values[self.gui_circular_crop]:
+        #     self.config.gui_circular_crop = values[self.gui_circular_crop]
         #    changed = True
 
         if changed:
@@ -248,16 +308,15 @@ class CameraWidget:
             # Event for mouse button up in ROI mode
             self.is_mouse_up = True
             if self.x1 < 0:
-                    self.x1 = 0
+                self.x1 = 0
             if self.y1 < 0:
-                    self.y1 = 0 
+                self.y1 = 0
             if abs(self.x0 - self.x1) != 0 and abs(self.y0 - self.y1) != 0:
                 self.config.roi_window_x = min([self.x0, self.x1])
                 self.config.roi_window_y = min([self.y0, self.y1])
                 self.config.roi_window_w = abs(self.x0 - self.x1)
                 self.config.roi_window_h = abs(self.y0 - self.y1)
                 self.main_config.save()
-                
 
         if event == self.gui_roi_selection:
             # Event for mouse button down or mouse drag in ROI mode
@@ -269,7 +328,7 @@ class CameraWidget:
         if event == self.gui_restart_calibration:
             self.ransac.calibration_frame_counter = self.settings.calibration_samples
             self.ransac.ibo.clear_filter()
-            PlaySound(resource_path('Audio/start.wav'), SND_FILENAME | SND_ASYNC)
+            PlaySound(resource_path("Audio/start.wav"), SND_FILENAME | SND_ASYNC)
 
         if event == self.gui_stop_calibration:
             self.ransac.calibration_frame_counter = 0
@@ -280,8 +339,8 @@ class CameraWidget:
         needs_roi_set = self.config.roi_window_h <= 0 or self.config.roi_window_w <= 0
 
         # TODO: Refactor if statements below...
-        window[self.gui_tracking_fps].update('')
-        window[self.gui_tracking_bps].update('')
+        window[self.gui_tracking_fps].update("")
+        window[self.gui_tracking_bps].update("")
         if self.config.capture_source is None or self.config.capture_source == "":
             window[self.gui_mode_readout].update("Waiting for camera address")
             window[self.gui_roi_message].update(visible=False)
@@ -329,6 +388,7 @@ class CameraWidget:
                 window[self.gui_roi_message].update(visible=False)
                 window[self.gui_output_graph].update(visible=True)
                 (maybe_image, eye_info) = self.image_queue.get(block=False)
+
                 imgbytes = cv2.imencode(".ppm", maybe_image)[1].tobytes()
                 window[self.gui_tracking_image].update(data=imgbytes)
 
@@ -336,13 +396,15 @@ class CameraWidget:
                 graph = window[self.gui_output_graph]
                 graph.erase()
 
-                if eye_info.info_type != EyeInfoOrigin.FAILURE: #and not eye_info.blink:
+                if (
+                    eye_info.info_type != EyeInfoOrigin.FAILURE
+                ):  # and not eye_info.blink:
                     graph.update(background_color="white")
                     if not np.isnan(eye_info.x) and not np.isnan(eye_info.y):
-                        
+
                         graph.draw_circle(
                             (eye_info.x * -100, eye_info.y * -100),
-                            20,
+                            eye_info.pupil_dilation * 25,
                             fill_color="black",
                             line_color="white",
                         )
@@ -356,12 +418,20 @@ class CameraWidget:
 
                     if not np.isnan(eye_info.blink):
 
-                        graph.draw_line((-100, eye_info.blink * 2 * 200), (-100, 100),  color="#6f4ca1", width=10)
+                        graph.draw_line(
+                            (-100, abs(eye_info.blink) * 2 * 200),
+                            (-100, 100),
+                            color="#6f4ca1",
+                            width=10,
+                        )
                     else:
-                        graph.draw_line((-100, 0.5 * 200), (-100, 100), color="#6f4ca1", width=10)
+                        graph.draw_line(
+                            (-100, 0.5 * 200), (-100, 100), color="#6f4ca1", width=10
+                        )
 
                     if eye_info.blink <= 0.0:
                         graph.update(background_color="#6f4ca1")
+
                 elif eye_info.info_type == EyeInfoOrigin.FAILURE:
                     graph.update(background_color="red")
                 # Relay information to OSC